+++ internal/repository/transactions.go
@@ -22,8 +22,8 @@ const (
 type TransactionsRepository interface {
 	Create(transaction *domain.Transaction) (*domain.Transaction, error)
 	UpdateByMsgID(transaction *domain.Transaction) (*domain.Transaction, error)
-	GetAll(userID int, filters *domain.TransactionFilters) (*[]domain.Transaction, error)
-	GetByMsgID(msgID int, userID int) (*domain.Transaction, error)
+	GetAll(userID int64, filters *domain.TransactionFilters) (*[]domain.Transaction, error)
+	GetByMsgID(msgID int, userID int64) (*domain.Transaction, error)
 }
 
 type transactionsRepository struct {
@@ -63,7 +63,7 @@ func (r *transactionsRepository) UpdateByMsgID(transaction *domain.Transaction) 
 	}
 	return transaction, nil
 }
-func (r *transactionsRepository) GetAll(userID int, filters *domain.TransactionFilters) (*[]domain.Transaction, error) {
+func (r *transactionsRepository) GetAll(userID int64, filters *domain.TransactionFilters) (*[]domain.Transaction, error) {
 	query, args, err := r.sqlBuilder.GetAllSQL(userID, filters)
 
 	rows, err := r.db.Queryx(query, args...)
@@ -85,7 +85,7 @@ func (r *transactionsRepository) GetAll(userID int, filters *domain.TransactionF
 
 	return &results, nil
 }
-func (r *transactionsRepository) GetByMsgID(msgID int, userID int) (*domain.Transaction, error) {
+func (r *transactionsRepository) GetByMsgID(msgID int, userID int64) (*domain.Transaction, error) {
 	query, args, err := r.sqlBuilder.GetByMsgIDSQL(msgID, userID)
 	if err != nil {
 		return nil, jsend.NewError("failed GetByMsgIDSQL", err, http.StatusInternalServerError)
@@ -125,7 +125,7 @@ func (tsql *transactionsSQL) UpdateByMsgIDSQL(transaction *domain.Transaction) (
 		PlaceholderFormat(sq.Dollar).
 		ToSql()
 }
-func (tsql *transactionsSQL) GetAllSQL(userID int, filters *domain.TransactionFilters) (string, []interface{}, error) {
+func (tsql *transactionsSQL) GetAllSQL(userID int64, filters *domain.TransactionFilters) (string, []interface{}, error) {
 	q := sq.Select("*").
 		From(tableTransactions)
 
@@ -142,7 +142,7 @@ func (tsql *transactionsSQL) GetAllSQL(userID int, filters *domain.TransactionFi
 	return q.PlaceholderFormat(sq.Dollar).
 		ToSql()
 }
-func (tsql *transactionsSQL) GetByMsgIDSQL(msgID int, userID int) (string, []interface{}, error) {
+func (tsql *transactionsSQL) GetByMsgIDSQL(msgID int, userID int64) (string, []interface{}, error) {
 	return sq.Select("*").
 		From(tableTransactions).
 		Where(sq.And{
